+++ packages/nice-grpc-client-middleware-deadline/src/index.ts
@@ -21,11 +21,12 @@ export const deadlineMiddleware: ClientMiddleware<DeadlineOptions> =
     origSignal?.addEventListener('abort', abortListener);
 
     let timedOut = false;
-    
-    const offset = deadline instanceof Date ? deadline.getTime() - Date.now() : deadline;
+
+    const offset =
+      deadline instanceof Date ? deadline.getTime() - Date.now() : deadline;
     const timer = setTimeout(() => {
       timedOut = true;
-      abortController.abort();      
+      abortController.abort();
     }, offset);
 
     try {