+++ packages/nice-grpc-client-middleware-deadline/CHANGELOG.md
@@ -4,10 +4,11 @@
 
 ## [1.1.0](https://github.com/deeplay-io/nice-grpc/compare/nice-grpc-client-middleware-deadline@1.0.11...nice-grpc-client-middleware-deadline@1.1.0) (2022-05-06)
 
-
 ### Features
 
-* **nice-grpc-client-middleware-deadline:** allow deadline as relative offset ([#130](https://github.com/deeplay-io/nice-grpc/issues/130)) ([7332b29](https://github.com/deeplay-io/nice-grpc/commit/7332b29869785abebc0217aa9426cabc55d0ad37))
+- **nice-grpc-client-middleware-deadline:** allow deadline as relative offset
+  ([#130](https://github.com/deeplay-io/nice-grpc/issues/130))
+  ([7332b29](https://github.com/deeplay-io/nice-grpc/commit/7332b29869785abebc0217aa9426cabc55d0ad37))
 
 ## 1.0.11 (2022-05-04)
 
@@ -32,4 +33,3 @@
 ## 1.0.1 (2021-06-29)
 
 ## 1.0.0 (2021-06-29)
-