+++ packages/nice-grpc-client-middleware-retry/CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ## 1.1.0 (2022-05-04)
 
-
 ### Features
 
-* add support for `ts-proto` server and client interfaces ([#126](https://github.com/deeplay-io/nice-grpc/issues/126)) ([67f7eb6](https://github.com/deeplay-io/nice-grpc/commit/67f7eb613455426d6b63a4027132060a8a572f65)), closes [#115](https://github.com/deeplay-io/nice-grpc/issues/115)
+- add support for `ts-proto` server and client interfaces
+  ([#126](https://github.com/deeplay-io/nice-grpc/issues/126))
+  ([67f7eb6](https://github.com/deeplay-io/nice-grpc/commit/67f7eb613455426d6b63a4027132060a8a572f65)),
+  closes [#115](https://github.com/deeplay-io/nice-grpc/issues/115)
 
 ## 1.0.7 (2022-04-18)
 
@@ -22,4 +24,3 @@
 ## 1.0.2 (2022-01-16)
 
 ## 1.0.1 (2021-09-21)
-